Centralize DTOs and enums in Shared project

+++ BlazorWasmClient.Shared/DTOs/TruckDto.cs
@@ -10,7 +10,7 @@ namespace BlazorWasmClient.Shared.DTOs
         public string LicensePlate { get; set; }
 
         [Required]
-        public decimal ClaimedRawMaterialWeight { get; set; }
+        public int ClaimedRawMaterialWeight { get; set; }
 
         [Required]
         [Range(0, double.MaxValue, ErrorMessage = "Weight must be a positive number")]
